Refactor/ephemeral store templates golden layout settings (#7622)
# Description

Moved some parts of `ephemeral.js` into separate files. The high line
count is from just moving things around, I did not add any new
functionality.


Fixes: #7620

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------
